pyproject.toml
[tool.poetry]
name = "genesis"
version = "0.1.0"
description = ""
authors = ["Alexander <job.yakovlev.alexander@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
uwsgi = "^2.0.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
